Minor bug fixing
Mortar shells take a long time to come down, need to fix

Bug with how the enemies attack, if its too far from them then they just dont attack because the waypoint is on another path.
Fix: raycast a line to check if the point is on a different path, if it is then get the difference of the attack point from the current position and move the attack point back to the current path.